# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Open Agents HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/agents/{agent}/install:
    get:
      operationId: getAgentInstallStatus
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallJob'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Get the current or last install job for one agent harness
      tags:
      - agents
    post:
      operationId: startAgentInstall
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAgentInstallRequest'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallJob'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Start an asynchronous install for one fixed agent harness
      tags:
      - agents
...
  /api/v1/agents/{agent}/verify:
    post:
      operationId: verifyAgentInstall
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallJob'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Verify an installed harness without reinstalling or probing authentication
      tags:
      - agents
...
  /api/v1/agents/install-jobs:
    get:
      operationId: listAgentInstallJobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentInstallJobsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return the latest durable install job for every agent harness
      tags:
      - agents
  /api/v1/agents/installers:
    get:
      operationId: listAgentInstallers
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentInstallerCatalogResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Resolve the safe installation plan for every supported agent harness
      tags:
      - agents
...
  /api/v1/settings:
    get:
      operationId: getSettings
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Read the daemon-owned user preferences
      tags:
      - settings
  /api/v1/settings/session-interface:
    patch:
      operationId: updateSessionInterface
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionInterfaceRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Choose the default interface for new sessions
      tags:
      - settings
...
  /api/v1/system/install/{target}:
    get:
      operationId: getSystemInstallStatus
      parameters:
      - description: 'Install target identifier: tmux, gh, opencode, or cloudflared.'
        in: path
        name: target
        required: true
        schema:
          description: 'Install target identifier: tmux, gh, opencode, or cloudflared.'
          enum:
          - tmux
          - gh
          - opencode
          - cloudflared
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallJob'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Get the current or last known install job status for a system target
      tags:
      - system
    post:
      operationId: startSystemInstall
      parameters:
      - description: 'Install target identifier: tmux, gh, opencode, or cloudflared.'
        in: path
        name: target
        required: true
        schema:
          description: 'Install target identifier: tmux, gh, opencode, or cloudflared.'
          enum:
          - tmux
          - gh
          - opencode
          - cloudflared
          type: string
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallJob'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Start (or return the already-running) install job for a fixed system
        target
      tags:
      - system
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AgentInstallJobsResponse:
      properties:
        jobs:
          items:
            $ref: '#/components/schemas/InstallJob'
          type: array
      required:
      - jobs
      type: object
    AgentInstallMethod:
      properties:
        available:
          type: boolean
        command:
          type: string
        expectedDestination:
          type: string
        id:
          type: string
        label:
          type: string
        reason:
          type: string
        recommended:
          type: boolean
        reinstallAvailable:
          type: boolean
        reinstallCommand:
          type: string
        reinstallReason:
          type: string
      required:
      - id
      - label
      - available
      - recommended
      - reinstallAvailable
      type: object
    AgentInstallPlan:
      properties:
        agentId:
          type: string
        automatic:
          type: boolean
        available:
          type: boolean
        command:
          type: string
        documentationUrl:
          type: string
        expectedDestination:
          type: string
        method:
          type: string
        methods:
          items:
            $ref: '#/components/schemas/AgentInstallMethod'
          type: array
        reason:
          type: string
      required:
      - agentId
      - available
      - automatic
      - method
      - documentationUrl
      - methods
      type: object
...
    AgentInstallerCatalogResponse:
      properties:
        agents:
          items:
            $ref: '#/components/schemas/AgentInstallPlan'
          type: array
      required:
      - agents
      type: object
...
    InstallJob:
      properties:
        command:
          description: Human-readable install command, e.g. "brew install tmux", for
            display even before/without output.
          type: string
        error:
          description: 'Set on failure or when the target is unsupported on this machine:
            the exec error, the Unsupported reason, or a timeout message.'
          type: string
        expectedDestination:
          description: Expected or adapter-resolved executable destination.
          type: string
        finishedAt:
          description: Absent until the job finishes.
          format: date-time
          type:
          - "null"
          - string
        method:
          description: Server-owned installation method selected for this harness
            job.
          type: string
        output:
          description: Combined stdout+stderr from the install command, tail-capped
            to the last ~4000 bytes.
          type: string
        startedAt:
          format: date-time
          type:
          - "null"
          - string
        status:
          description: Current lifecycle state of the job.
          enum:
          - idle
          - running
          - installing
          - verifying
          - succeeded
          - failed
          - unsupported
          - interrupted
          type: string
        target:
          description: Fixed install target this job ran (or is running) for.
          enum:
          - tmux
          - gh
          - opencode
          - cloudflared
          type: string
        updatedAt:
          format: date-time
          type:
          - "null"
          - string
      required:
      - target
      - status
      type: object
...
    SettingsResponse:
      properties:
        chatHarnesses:
          items:
            type: string
          type: array
        defaultSessionMode:
          enum:
          - chat
          - tui
          type: string
      required:
      - defaultSessionMode
      - chatHarnesses
      type: object
...
    StartAgentInstallRequest:
      properties:
        method:
          description: Server-issued installation method id. Omit to use the recommended
            viable method.
          type: string
        operation:
          description: Requested operation. Defaults to install for older clients.
          enum:
          - install
          - reinstall
          type: string
      type: object
...
    UpdateSessionInterfaceRequest:
      properties:
        defaultSessionMode:
          enum:
          - chat
          - tui
          type: string
      required:
      - defaultSessionMode
      type: object
...